Run bundle update to fix deprecated/incompatible gems
This is a workaround to fix this error
`your bundle is locked to mimemagic (0.3.5) from rubygems repository
https://rubygems.org/, cached gems or installed locally, but that
version can no longer be found in that source. that means the author of
mimemagic (0.3.5) has removed it. you'll need to update your bundle to a
version other than mimemagic (0.3.5) that hasn't been removed in order
to install.`

P.S. since this is a timed task I did not want to spend too much time
trying to fix this. It should be noted that in a work environment I
would have dedicated some time to investigate furthur and carefully
update only the necessary gems