Challenge 1: Add review form and most styling
Adds the review form and a spec to ensure the correct product is
reviewed.

Most of the styling is included in this commit with the notable
exclusion of stars. Those are more complicated as realisatically they
will require the addition of an icon font. However, they're not top
priority.

I've added Factory Bot to the app to make generating records easier in
specs/tests.

The async back end will come in a subsequent commit.