Challenge 1: Create reviews asynchronously
This replaces the old, synchronous create code with a Sidekiq-based
asynchronous solution. In addition to improving performance for the user
by deferring the work, this has the added bonus of taking that code out
of the controller to somewhere more suitable.

In refactoring this code, I removed the incomplete feature which
theoretically allowed reviewers to submit their own tags. Our front end
does not support this and currently there are no plans to change that,
so it doesn't make sense to hold onto this.