Challenge 2: Serve styled, paginated reviews on index page
I've revamped the index page with styling and pagination. To deliver an
MVP, the pagination is server-side for the time being. No pagination
exists for reviews of products.

Whilst not ideal, server-side works as a trade-off in this case because
Turbolinks speeds up navigation with implicit AJAX/XHR loading anyway.

A barebones system test has been written for this page.